Extracted test cases for 'Sets' strategy into separate modules
